refactor: extract people/email read queries into shared lib modules
Moves the read-query logic behind GET /api/people, /api/people/:id,
/api/emails/by-person/:personId, /api/emails/:id and PATCH /api/emails/:id
into worker/src/lib/queries/{people,emails}.ts.

The MCP server needs this logic too. Reimplementing it there is what the
removed MCP router did — 646 lines duplicating router queries, guaranteed to
drift from the HTTP surface. Both surfaces now call one implementation.

Every function takes `allowed: AllowedInboxes` as an explicit parameter rather
than reading Hono context, which is what makes them callable from a transport
with no request context. That is the pattern emails-router already used with
inboxFilter; it is now the rule for this layer.

Pure refactor. Status codes and response bodies are unchanged, including the
existing behavior of returning 404 rather than 403 for a permission denial, so
callers can't probe for ids they can't see. Routers keep their OpenAPI
definitions and response shaping and delegate only the data access.

parseCc moves into lib/queries/emails.ts because the extracted code needs it
and conversations-router did too; importing it back from the router would have
created a router -> lib -> router cycle.

Verified: yarn tsc --noEmit clean, full suite 63 files / 496 passed.
