+++ worker/src/routers/people-router.ts
@@ -5,21 +5,12 @@ import { emails } from "../db/emails.schema";
 import { attachments } from "../db/attachments.schema";
 import { sentEmails } from "../db/sent-emails.schema";
 import { json200Response, escapeLike, escapeFts } from "../lib/helpers";
+import {
+  getPersonScoped,
+  listPeople,
+  peopleScopeClause,
+} from "../lib/queries/people";
 import type { Variables } from "../variables";
-import type { AllowedInboxes } from "../lib/inbox-permissions";
-
-function peopleScopeClause(allowed: AllowedInboxes) {
-  if (allowed.isAdmin) return sql``;
-  if (allowed.inboxes.length === 0)
-    return sql`AND s.id IN (SELECT NULL WHERE 0)`;
-  // A person is in scope if they emailed one of our allowed inboxes OR if we
-  // sent them mail from one of our allowed inboxes.
-  return sql`AND s.id IN (
-    SELECT person_id FROM emails WHERE recipient IN ${allowed.inboxes}
-    UNION
-    SELECT person_id FROM sent_emails WHERE from_address IN ${allowed.inboxes} AND person_id IS NOT NULL
-  )`;
-}
 
 export const peopleRouter = new OpenAPIHono<{
   Bindings: CloudflareBindings;
@@ -596,78 +587,15 @@ const listPeopleRoute = createRoute({
 peopleRouter.openapi(listPeopleRoute, async (c) => {
   const db = c.get("db");
   const { q, recipient, personId, page, limit } = c.req.valid("query");
-  const offset = (page - 1) * limit;
-
-  // Build WHERE conditions for the emails table
-  const conditions: any[] = [];
-
-  if (q) {
-    const pattern = `%${escapeLike(q)}%`;
-    conditions.push(
-      sql`(s.email LIKE ${pattern} ESCAPE '\\' OR s.name LIKE ${pattern} ESCAPE '\\')`,
-    );
-  }
-
-  if (recipient) {
-    conditions.push(sql`e.recipient = ${recipient}`);
-  }
-
-  if (personId) {
-    conditions.push(sql`s.id = ${personId}`);
-  }
-
   const allowed = c.get("allowedInboxes")!;
-  const scopeClause = peopleScopeClause(allowed);
-  const extraConditions =
-    conditions.length > 0
-      ? sql`AND ${sql.join(conditions, sql` AND `)}`
-      : sql``;
-  const whereClause = sql`WHERE 1=1 ${extraConditions} ${scopeClause}`;
 
-  // Group by (person, recipient) to get per-thread stats
-  const rows = await db.all<{
-    id: string;
-    email: string;
-    name: string | null;
-    recipient: string;
-    lastEmailAt: number;
-    unreadCount: number;
-    totalCount: number;
-    latestSubject: string | null;
-  }>(sql`
-    SELECT
-      s.id,
-      s.email,
-      s.name,
-      e.recipient,
-      MAX(e.received_at) AS lastEmailAt,
-      SUM(CASE WHEN e.is_read = 0 THEN 1 ELSE 0 END) AS unreadCount,
-      COUNT(*) AS totalCount,
-      (
-        SELECT e2.subject FROM emails e2
-        WHERE e2.person_id = s.id AND e2.recipient = e.recipient
-        ORDER BY e2.received_at DESC LIMIT 1
-      ) AS latestSubject
-    FROM ${emails} e
-    JOIN ${people} s ON s.id = e.person_id
-    ${whereClause}
-    GROUP BY s.id, e.recipient
-    ORDER BY lastEmailAt DESC
-    LIMIT ${limit} OFFSET ${offset}
-  `);
+  const result = await listPeople(
+    db,
+    { q, recipient, personId, page, limit },
+    allowed,
+  );
 
-  // Get total count of (person, recipient) pairs
-  const countResult = await db.all<{ count: number }>(sql`
-    SELECT COUNT(*) AS count FROM (
-      SELECT 1 FROM ${emails} e
-      JOIN ${people} s ON s.id = e.person_id
-      ${whereClause}
-      GROUP BY s.id, e.recipient
-    )
-  `);
-  const total = countResult[0]?.count ?? 0;
-
-  return c.json({ data: rows, total, page, limit }, 200);
+  return c.json(result, 200);
 });
 
 const getPersonRoute = createRoute({
@@ -689,33 +617,14 @@ peopleRouter.openapi(getPersonRoute, async (c) => {
   const db = c.get("db");
   const { id } = c.req.valid("param");
 
-  const rows = await db.select().from(people).where(eq(people.id, id)).limit(1);
+  const allowed = c.get("allowedInboxes")!;
+  const person = await getPersonScoped(db, id, allowed);
 
-  if (rows.length === 0) {
+  if (!person) {
     return c.json({ error: "Person not found" }, 404);
   }
 
-  const allowed = c.get("allowedInboxes")!;
-  if (!allowed.isAdmin) {
-    if (allowed.inboxes.length === 0) {
-      return c.json({ error: "Person not found" }, 404);
-    }
-    const match = await db
-      .select({ id: emails.id })
-      .from(emails)
-      .where(
-        and(
-          eq(emails.personId, id),
-          inArray(emails.recipient, allowed.inboxes),
-        ),
-      )
-      .limit(1);
-    if (match.length === 0) {
-      return c.json({ error: "Person not found" }, 404);
-    }
-  }
-
-  return c.json(rows[0], 200);
+  return c.json(person, 200);
 });
 
 const deletePersonRoute = createRoute({
